requirements.txt
Django==3.1.4
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pymongo==3.11.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn-pandas==2.0.3
sqlparse==0.4.1
threadpoolctl==2.1.0
Werkzeug==1.0.1
whitenoise==5.2.0
wincertstore==0.2
